requirements.txt
allure-pytest==2.8.36
allure-python-commons==2.8.36
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==4.4.2
idna==2.10
importlib-metadata==3.7.2
iniconfig==1.1.1
jsonpath-rw==1.4.0
packaging==20.9
pluggy==0.13.1
ply==3.11
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-ordering==0.6
requests==2.25.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
xlrd==1.2.0
zipp==3.4.1
pymysql==1.0.2
